@@ -4,8 +4,8 @@ import 'package:gametest/bunker-game.dart';
 
 class Skybox {
   final BunkerGame game;
-  Sprite bgSprite;
-  Rect bgRect;
+  late Sprite bgSprite;
+  late Rect bgRect;
 
   Skybox(this.game) {
 
